// frontend/texttosql/src/app/api/models/message-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MessageDto {
  chatId?: string;
  confidence?: number;
  content: string;
  createdAt?: string;
  databaseId?: string;
  feedback?: 'GOOD' | 'BAD' | 'NONE';
  messageId?: string;
  senderType?: 'LLM' | 'USER';
}